requirements.txt
discord.py
disputils
python-decouple
